requirements.txt
# Core Azure Authentication & SDK Base
azure-identity>=1.12.0
msal

# Azure Management Libraries (Fetchers)
azure-mgmt-resource>=21.1.0
azure-mgmt-network>=19.1.0
azure-mgmt-compute>=27.2.0
azure-mgmt-storage>=19.1.0
azure-mgmt-web>=7.0.0
azure-mgmt-keyvault>=9.3.0
azure-mgmt-authorization>=3.0.0
azure-mgmt-policyinsights
azure-mgmt-security>=2.0.0
azure-mgmt-advisor>=9.0.0
azure-mgmt-consumption>=9.0.0
azure-mgmt-costmanagement>=3.0.0
azure-mgmt-loganalytics>=12.0.0
azure-mgmt-privatedns
azure-mgmt-cosmosdb>=9.2.0
azure-mgmt-managementgroups
azure-mgmt-monitor>=5.0.0
azure-mgmt-sql>=3.0.0
azure-keyvault-secrets
azure-keyvault-certificates>=4.7.0
msgraph-sdk

# Azure AI Services
azure-mgmt-cognitiveservices>=13.5.0
azure-mgmt-machinelearningservices>=2.0.0b2
azure-mgmt-search>=9.0.0
azure-mgmt-botservice>=2.0.0

# Report/Diagram Generation
markdown>=3.4.1
graphviz>=0.20.1
jinja2

# Optional Features & Utilities
openai
python-dotenv>=0.19.2
rich>=13.3.1
mdutils>=1.3.1
requests>=2.28.1